fix some bosminer bugs with new make types and graphql, and rename `models.py` file to `makes.py` to be consistent with naming conventions.
