+++ pyasic/miners/_types/antminer/X17/S17.py
@@ -12,7 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from pyasic.miners._types.models import AntMiner
+from pyasic.miners._types.makes import AntMiner
 
 class S17(AntMiner):  # noqa - ignore ABC method implementation
     def __init__(self, ip: str):
