+++ pyasic/miners/_types/antminer/X19/S19a.py
@@ -12,7 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from pyasic.miners._types.models import AntMiner
+from pyasic.miners._types.makes import AntMiner
 
 class S19a(AntMiner):  # noqa - ignore ABC method implementation
     def __init__(self, ip: str):
