+++ pyasic/miners/_types/avalonminer/A10X/A1026.py
@@ -12,7 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from pyasic.miners._types.models import AvalonMiner
+from pyasic.miners._types.makes import AvalonMiner
 
 class Avalon1026(AvalonMiner):  # noqa - ignore ABC method implementation
     def __init__(self, ip: str):
