+++ pyasic/miners/_types/whatsminer/M2X/M21.py
@@ -12,7 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from pyasic.miners._types.models import WhatsMiner
+from pyasic.miners._types.makes import WhatsMiner
 
 class M21(WhatsMiner):  # noqa - ignore ABC method implementation
     def __init__(self, ip: str):
