+++ pyasic/miners/_types/whatsminer/M3X/M30S_Plus_Plus.py
@@ -12,7 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from pyasic.miners._types.models import WhatsMiner
+from pyasic.miners._types.makes import WhatsMiner
 
 class M30SPlusPlus(WhatsMiner):  # noqa - ignore ABC method implementation
     def __init__(self, ip: str):
